Refactor coordinate serialization and deserialization
This commit refactors the serialization and deserialization logic for coordinates in the GTFS format. It introduces a new module `coord_type` as an enum workaround for the lack of support for associated enum constants in Rust. The `GtfsCoord` struct is now generic and takes a coordinate type as a type parameter. The `GtfsStopCoordFlatten` and `GtfsShapeCoordFlatten` structs are introduced to represent GTFS coordinates for stops and shapes respectively. The `GtfsCoordFlatten` struct is removed. The `GtfsCoord` struct now implements the `Serialize` and `Deserialize` traits based on the coordinate type. The `deserialize` function now uses a custom visitor based on the coordinate type. The `parse_float` function is moved to the `common.rs` file. The `GtfsCoord` struct now dereferences to the `Coord` struct. The `GtfsCoord` struct now implements `Deref` and `DerefMut` traits. The `stop_coord` field in the `Stop` struct now uses the `GtfsCoord` type with the `coord_type::STOP` constant. The `shape_pt` field in the `Shape` struct now uses the `GtfsCoord` type with the `coord_type::SHAPE` constant.